+++ kbstatpy/__init__.py
@@ -1,5 +1,5 @@
 from .options import KbstatOptions
 from .kbstat import Kbstat
 
-__version__ = "1.13.4"
+__version__ = "1.13.5"
 __all__ = ["Kbstat", "KbstatOptions", "__version__"]